requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
openai==1.3.5
redis==5.0.1
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3
PyGithub==2.1.1
onedrivesdk==2.0.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.114.0
cryptography==42.0.1
regex==2023.12.25 